refactor: trace utils
