+++ src/ape/utils/misc.py
@@ -29,6 +29,7 @@ _python_version = (
     f"{sys.version_info.major}.{sys.version_info.minor}"
     f".{sys.version_info.micro} {sys.version_info.releaselevel}"
 )
+ZERO_ADDRESS = "0x0000000000000000000000000000000000000000"
 
 
 @lru_cache(maxsize=None)
